+++ packages/eslint-plugin/typings/ts-eslint.d.ts
@@ -475,14 +475,24 @@ declare module 'ts-eslint' {
     Token?: RuleFunction<TSESTree.Token>;
     TryStatement?: RuleFunction<TSESTree.TryStatement>;
     TSAbstractKeyword?: RuleFunction<TSESTree.TSAbstractKeyword>;
+    TSAbstractMethodDefinition?: RuleFunction<
+      TSESTree.TSAbstractMethodDefinition
+    >;
     TSAnyKeyword?: RuleFunction<TSESTree.TSAnyKeyword>;
     TSArrayType?: RuleFunction<TSESTree.TSArrayType>;
     TSAsExpression?: RuleFunction<TSESTree.TSAsExpression>;
     TSAsyncKeyword?: RuleFunction<TSESTree.TSAsyncKeyword>;
     TSBigIntKeyword?: RuleFunction<TSESTree.TSBigIntKeyword>;
     TSBooleanKeyword?: RuleFunction<TSESTree.TSBooleanKeyword>;
+    TSCallSignatureDeclaration?: RuleFunction<
+      TSESTree.TSCallSignatureDeclaration
+    >;
     TSConditionalType?: RuleFunction<TSESTree.TSConditionalType>;
+    TSConstructSignatureDeclaration?: RuleFunction<
+      TSESTree.TSConstructSignatureDeclaration
+    >;
     TSDeclareKeyword?: RuleFunction<TSESTree.TSDeclareKeyword>;
+    TSDeclareFunction?: RuleFunction<TSESTree.TSDeclareFunction>;
     TSEnumDeclaration?: RuleFunction<TSESTree.TSEnumDeclaration>;
     TSEnumMember?: RuleFunction<TSESTree.TSEnumMember>;
     TSExportAssignment?: RuleFunction<TSESTree.TSExportAssignment>;