+++ packages/typescript-estree/tests/ast-alignment/fixtures-to-test.ts
@@ -389,6 +389,12 @@ tester.addFixturePatternConfig('typescript/basics', {
      * TODO: remove me in next babel > 7.3.2
      */
     'arrow-function-with-optional-parameter',
+    /**
+     * [BABEL ERRORED, BUT TS-ESTREE DID NOT]
+     */
+    'const-assertions',
+    'readonly-arrays',
+    'readonly-tuples',
   ],
   ignoreSourceType: [
     /**